third_party/blink/web_tests/svg/dynamic-updates/SVGFEDisplacementMapElement-svgdom-in2-prop-expected.txt

SVG 1.1 dynamic update tests

Tests dynamic updates of the 'in2' property of the SVGFEDisplacementMapElement object

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS displacementMap.in2.baseVal is "Texture"
PASS displacementMap.in2.baseVal is "Map"
PASS successfullyParsed is true

TEST COMPLETE
